requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-jose==3.3.0
bcrypt==4.3.0
python-multipart==0.0.6
pydantic>=2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0
cryptography==41.0.7
requests==2.31.0
httpx==0.25.2
scikit-learn==1.3.2
jieba==0.42.1
oss2==2.18.1
langchain-core==0.3.29
langchain-openai==0.3.0
numpy==1.26.2
